[feat] DDD Layered Architecture 구조 변경
