+++ hub/src/main/java/com/springcloud/hub/domain/api/NaverMapClient.java
@@ -1,0 +1,7 @@
+package com.springcloud.hub.domain.api;
+
+import java.math.BigDecimal;
+
+public interface NaverMapClient {
+    String requestOptimalRoute(BigDecimal startLat, BigDecimal startLon, BigDecimal goalLat, BigDecimal goalLon);
+}
