+++ hub/src/main/java/com/springcloud/hub/domain/repository/HubRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.springcloud.hub.infrastructure;
+package com.springcloud.hub.domain.repository;
 
 import com.springcloud.hub.domain.entity.Hub;
