+++ hub/src/main/java/com/springcloud/hub/interfaces/GetHubResponseDto.java
@@ -1,4 +1,4 @@
-package com.springcloud.hub.application;
+package com.springcloud.hub.interfaces;
 
 import com.springcloud.hub.domain.entity.Hub;
